[Lesson 205]

+++ includes/view_all_posts.php
@@ -110,7 +110,7 @@
             <a href="posts.php?source=edit_post&p_id=<?php echo $post_id; ?>">Edit</a>
           </td>
           <td>
-            <a href="posts.php?delete=<?php echo $post_id; ?>">Delete</a>
+            <a onClick="javascript: return confirm('Are you sure you want to delete?');" href="posts.php?delete=<?php echo $post_id; ?>">Delete</a>
           </td>
         </tr>
       <?php } ?>
